Attempt to stop locked files causing ouroboroi
If all the files in the folder were locked or had other permission issues (e.g. were created by another on a shared drive with some odd permission settings), the code would stick on checking (status=2) forever. Ultimately, the IsFileLocked function is too sensitive to be used and is in any case unnecessary because the code always treats any files as new.

Currently the old code is still present but is unused; until proper testing has been done I shan't remove it entirely.